+++ frontend/src/components/Datasets/Datasets.vue
@@ -26,7 +26,7 @@
         </v-row>
 
         <v-row class="bordered">
-            <v-col cols=4>Filter By:</v-col>
+            <v-col cols=4>Search/Filter By:</v-col>
             <v-col cols=8></v-col>
             <v-col cols=6>
                 <Select
@@ -35,18 +35,26 @@
                     :label="$tc('Ministry / Organization')"
                     :multiple="true"
                     :items="filterOrgItems"
-                    :value="filterOrg"
+                    :value="filterOrg"                    
                     @edited=" (newVal) => { filterOrg = newVal; }">
                 </Select>
             </v-col>
             <v-col cols=6>
-                <TextInput
-                    :editing="true"
+                <Select
+                    :label="$tc('Datasets') + ' ' + $tc('Name')"
+                    placeholder=""
                     name="name"
-                    :label="$tc('Name')"
+                    :items="repoNames"
+                    itemText="name"
+                    itemValue="name"
+                    :sorted="false"
+                    :combobox="true"
+
+                    :editing="true"
                     :value="filterName"
-                    @blur=" (newVal) => { filterName = newVal.target.value; }">
-                </TextInput>
+                    @edited="(newVal) => { filterNameEdit(newVal) }"
+                    @blur="(newVal) => { filterNameBlur(newVal) }"
+                ></Select>
             </v-col>
             <v-col cols=6>
                 <Select
@@ -130,15 +138,16 @@
 import Vue from 'vue';
 import {mapActions, mapState, mapMutations} from "vuex";
 import Select from '../FormElements/Select';
-import TextInput from '../FormElements/TextInput';
+// import TextInput from '../FormElements/TextInput';
 import TableReport from '../Reports/TableReport';
 
 const taggedFieldsRep = 'Tagged fields';
+const allFieldsExp = "All fields";
 
 export default {
     components:{
         Select,
-        TextInput,
+        // TextInput,
         TableReport
     },
 
@@ -149,7 +158,7 @@ export default {
             filterOrg: [],
             filterName: '',
             isUpdating: false,
-            reports: [taggedFieldsRep],
+            reports: [taggedFieldsRep, allFieldsExp],
             reportHeaders: [],
             reportItems: [],
             showReport: false,
@@ -159,7 +168,8 @@ export default {
             datasetTags: {},
             sortBy: "create_date desc",
             sortField: "create_date",
-            sortDir: 'desc'
+            sortDir: 'desc',
+            repoNames: []
         }
     },
     async mounted(){
@@ -174,11 +184,24 @@ export default {
         ...mapMutations({
             setSelectedFilterBy: 'repos/setSelectedFilterBy',
         }),
+
+        filterNameBlur(newVal){
+            this.filterName = newVal
+        },
+
+        filterNameEdit(newVal){
+            this.filterName = typeof(newVal) === 'string' ? newVal : newVal.name;
+        },
+
         async loadRepos() {
             this.message = this.$tc('Loading') + ' ' + this.$tc('Datasets', 2) + '...';
             
             await this.getRepos({filterBy: false});
             this.filteredRepos = this.repos;
+            this.repoNames = [{name: ''}];
+            for (let i=0; i<this.repos.length; i++){
+                this.repoNames.push({name: this.repos[i].name});
+            }
 
             this.getReposFull({});
             this.message = '';
@@ -221,6 +244,36 @@ export default {
 
         goReport(report){
             let items = [];
+            if (this.reposFull){
+                this.reposFull.forEach( (repo) => {
+                    let add = (this.filterOrg.length === 0 || this.filterOrg.indexOf(repo.version.repo_id.ministry_organization) !== -1);
+                    add = ( (add) && (!this.filterName || repo.version.repo_id.name.indexOf(this.filterName) !== -1) );
+                    
+                    if (add){
+                        for (let i=0; i<repo.resources.length; i++){
+                            if (repo.resources[i].tableSchema && repo.resources[i].tableSchema.fields){
+
+                                for (let j=0; j<repo.resources[i].tableSchema.fields.length; j++){
+                                    let item = JSON.parse(JSON.stringify(repo));
+                                    item.field = JSON.parse(JSON.stringify(repo.resources[i].tableSchema.fields[j]));
+                                    item.resource = JSON.parse(JSON.stringify(repo.resources[i]));
+                                    let tags = item.field.tags;
+                                    let intersection = [];
+                                    if (tags && this.filterTags && tags.filter){
+                                        intersection = tags.filter(element => this.filterTags.includes(element));
+                                    }
+                                    
+                                    if (!this.filterTags || (tags && intersection && intersection.length > 0)){
+                                        items.push(item);
+                                    }
+                                }
+                            }
+                        }
+                        
+                    }
+                });
+            }
+
             if (report === taggedFieldsRep){
                 this.reportHeaders = [
                     {
@@ -259,32 +312,299 @@ export default {
                         value: 'field.tags',
                     },
                 ];
-
-                if (this.reposFull){
-                    this.reposFull.forEach( (repo) => {
-                        let add = (this.filterOrg.length === 0 || this.filterOrg.indexOf(repo.version.repo_id.ministry_organization) !== -1);
-                        add = ( (add) && (!this.filterName || repo.version.repo_id.name.indexOf(this.filterName) !== -1) );
-                        
-                        if (add){
-                            for (let i=0; i<repo.resources.length; i++){
-                                if (repo.resources[i].tableSchema && repo.resources[i].tableSchema.fields){
-
-                                    for (let j=0; j<repo.resources[i].tableSchema.fields.length; j++){
-                                        let item = JSON.parse(JSON.stringify(repo));
-                                        item.field = JSON.parse(JSON.stringify(repo.resources[i].tableSchema.fields[j]));
-                                        item.resource = JSON.parse(JSON.stringify(repo.resources[i]));
-                                        let tags = item.field.tags;
-                                        
-                                        if (!this.filterTags || (tags && tags.indexOf(this.filterTags) !== -1)){
-                                            items.push(item);
-                                        }
-                                    }
-                                }
-                            }
-                            
-                        }
-                    });
-                }
+            }else if (report === allFieldsExp){
+                this.reportHeaders = [
+                    {
+                        text: 'Ministry/Organization',
+                        sortable: true,
+                        value: 'version.repo_id.ministry_organization',
+                    },
+                    {
+                        text: 'Dataset Name',
+                        sortable: true,
+                        value: 'version.repo_id.name',
+                    },
+                    {
+                        text: 'Dataset Description',
+                        sortable: false,
+                        value: 'version.repo_id.description',
+                    },
+                    {
+                        text: 'Sector',
+                        sortable: true,
+                        value: 'version.repo_id.sector',
+                    },
+                    {
+                        text: 'Data Type',
+                        sortable: false,
+                        value: 'version.repo_id.data_type',
+                    },
+                    {
+                        text: 'Data Collection Type',
+                        sortable: false,
+                        value: 'version.repo_id.data_collection_type',
+                    },
+                    {
+                        text: 'Restrictions Comments',
+                        sortable: false,
+                        value: 'version.repo_id.restrictions_comments',
+                    },
+                    {
+                        text: 'Allow Academic Publish',
+                        sortable: false,
+                        value: 'version.repo_id.aca_allow_publish',
+                    },
+                    {
+                        text: 'Academic Approval Needed',
+                        sortable: false,
+                        value: 'version.repo_id.aca_approval_needed',
+                    },
+                    {
+                        text: 'Allow Gov Publish',
+                        sortable: false,
+                        value: 'version.repo_id.gov_allow_publish',
+                    },
+                    {
+                        text: 'Gov Approval Needed',
+                        sortable: false,
+                        value: 'version.repo_id.gov_approval_needed',
+                    },
+                    {
+                        text: 'In BCDC',
+                        sortable: false,
+                        value: 'version.repo_id.in_bc_catalogue',
+                    },
+                    {
+                        text: 'Contact Info',
+                        sortable: false,
+                        value: 'version.repo_id.contact',
+                    },
+                    {
+                        text: 'Dataset Refresh Status',
+                        sortable: false,
+                        value: 'version.repo_id.refresh_status',
+                    },
+                    {
+                        text: 'Dataset Refresh Schedule',
+                        sortable: false,
+                        value: 'version.repo_id.refresh_schedule',
+                    },
+                    {
+                        text: 'Dataset Lifecycle Status',
+                        sortable: false,
+                        value: 'version.repo_id.lifecycle_status',
+                    },
+                    {
+                        text: 'Dataset Lifecycle Dates',
+                        sortable: false,
+                        value: 'version.repo_id.lifecycle_dates',
+                    },
+                    {
+                        text: 'Dataset Provider Groups',
+                        sortable: false,
+                        value: 'version.repo_id.author_groups',
+                    },
+                    {
+                        text: this.$tc('Version', 1) + ' Name',
+                        sortable: true,
+                        value: 'version.name',
+                    },
+                    {
+                        text: 'Short Name',
+                        sortable: true,
+                        value: 'version.short_title',
+                    },
+                    {
+                        text: this.$tc('Version', 1) + ' Notes',
+                        sortable: false,
+                        value: 'version.description',
+                    },
+                    {
+                        text: 'Keywords',
+                        sortable: false,
+                        value: 'version.keywords',
+                    },
+                    {
+                        text: this.$tc('Version', 1) + ' Type',
+                        sortable: false,
+                        value: 'version.type',
+                    },
+                    {
+                        text: 'Variable Classification Index',
+                        sortable: false,
+                        value: 'version.variable_classification',
+                    },
+                    {
+                        text: 'Citation',
+                        sortable: false,
+                        value: 'version.citation',
+                    },
+                    {
+                        text: this.$tc('Version', 1) + ' Lifecycle',
+                        sortable: false,
+                        value: 'version.lifecycle',
+                    },
+                    {
+                        text: 'Specific Instructions',
+                        sortable: false,
+                        value: 'version.instructions',
+                    },
+                    {
+                        text: 'Linking Results',
+                        sortable: false,
+                        value: 'version.linking_summary',
+                    },
+                    {
+                        text: 'Processing Summary',
+                        sortable: false,
+                        value: 'version.processing_summary',
+                    },
+                    {
+                        text: 'Collection Method',
+                        sortable: false,
+                        value: 'version.collectionMethod',
+                    },
+                    {
+                        text: 'Inclusions',
+                        sortable: false,
+                        value: 'version.inclusions',
+                    },
+                    {
+                        text: 'Exclusions',
+                        sortable: false,
+                        value: 'version.exclusions',
+                    },
+                    {
+                        text: 'Quality',
+                        sortable: false,
+                        value: 'version.quality',
+                    },
+                    {
+                        text: 'Data Changes Over Time',
+                        sortable: false,
+                        value: 'version.delta_over_time',
+                    },
+                    {
+                        text: 'Additional Info',
+                        sortable: false,
+                        value: 'version.additional_info',
+                    },
+                    {
+                        text: 'Links',
+                        sortable: false,
+                        value: 'version.more_information',
+                    },
+                    {
+                        text: 'References',
+                        sortable: false,
+                        value: 'version.references',
+                    },
+                    {
+                        text: 'Approved',
+                        sortable: false,
+                        value: 'version.approved',
+                    },
+                    {
+                        text: 'Published',
+                        sortable: false,
+                        value: 'version.published',
+                    },
+                    {
+                        text: 'FAQ',
+                        sortable: false,
+                        value: 'version.faq',
+                    },
+                    {
+                        text: this.$tc('Resource', 1) + ' Name',
+                        sortable: true,
+                        value: 'resource.name',
+                    },
+                    {
+                        text: this.$tc('Resource', 1) + ' Path',
+                        sortable: true,
+                        value: 'resource.path',
+                    },
+                    {
+                        text: this.$tc('Resource', 1) + ' Description',
+                        sortable: true,
+                        value: 'resource.description',
+                    },
+                    {
+                        text: this.$tc('Resource', 1) + ' Date Range Start',
+                        sortable: true,
+                        value: 'resource.temporal_start',
+                    },
+                    {
+                        text: this.$tc('Resource', 1) + ' Date Range End',
+                        sortable: true,
+                        value: 'resource.temporal_end',
+                    },
+                    {
+                        text: this.$tc('Resource', 1) + ' Source',
+                        sortable: true,
+                        value: 'resource.source_system',
+                    },
+                    {
+                        text: this.$tc('Field', 1) + ' Name',
+                        sortable: true,
+                        value: 'field.name',
+                    },
+                    {
+                        text: this.$tc('Field', 1) + ' Title',
+                        sortable: true,
+                        value: 'field.title',
+                    },
+                    {
+                        text: this.$tc('Field', 1) + ' Source Name',
+                        sortable: true,
+                        value: 'field.shortName',
+                    },
+                    {
+                        text: this.$tc('Field', 1) + ' Type',
+                        sortable: true,
+                        value: 'field.type',
+                    },
+                    {
+                        text: this.$tc('Field', 1) + ' Format',
+                        sortable: false,
+                        value: 'field.format',
+                    },
+                    {
+                        text: this.$tc('Field', 1) + ' Variable Classification',
+                        sortable: false,
+                        value: 'field.var_class',
+                    },
+                    {
+                        text: this.$tc('Field', 1) + ' RDF Type',
+                        sortable: false,
+                        value: 'field.rdfType',
+                    },
+                    {
+                        text: this.$tc('Field', 1) + this.$tc('Tags', 2),
+                        sortable: false,
+                        value: 'field.tags',
+                    },
+                    {
+                        text: this.$tc('Field') + ' ' + this.$tc('Notes', 2),
+                        sortable: false,
+                        value: 'field.notes',
+                    },
+                    {
+                        text: this.$tc('Field') + ' ' + this.$tc('Enum', 1),
+                        sortable: false,
+                        value: 'field.constraints.enum',
+                    },
+                    {
+                        text: this.$tc('Field') + ' ' + this.$tc('Highlight'),
+                        sortable: false,
+                        value: 'field.highlight',
+                    },
+                    {
+                        text: this.$tc('Field') + ' as ' + this.$tc('json'),
+                        sortable: false,
+                        value: 'field',
+                    }
+                ];
             }
             this.reportItems = items;
             this.reportTitle = report;