+++ frontend/src/components/pages/publishedVariableClassification.vue
@@ -45,6 +45,7 @@ export default {
 <style scoped>
     .fullWidth{
         width: 100%;
+        overflow: visible
     }
 
 </style>